Created exceptions name and id

+++ src/main/java/br/com/andrezasecon/b2w/apiplanet/controller/PlanetController.java
@@ -3,8 +3,11 @@ package br.com.andrezasecon.b2w.apiplanet.controller;
 import br.com.andrezasecon.b2w.apiplanet.client.ClientFeignSwapi;
 import br.com.andrezasecon.b2w.apiplanet.client.PlanetSwapiPaginationResponse;
 import br.com.andrezasecon.b2w.apiplanet.domain.Planet;
+import br.com.andrezasecon.b2w.apiplanet.exceptions.InvalidIdException;
+import br.com.andrezasecon.b2w.apiplanet.exceptions.InvalidNameException;
+import br.com.andrezasecon.b2w.apiplanet.exceptions.NotFoundNameException;
+import br.com.andrezasecon.b2w.apiplanet.exceptions.PlanetNotFoundException;
 import br.com.andrezasecon.b2w.apiplanet.service.PlanetService;
-import br.com.andrezasecon.b2w.apiplanet.service.PlanetServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -15,7 +18,7 @@ import javax.validation.Valid;
 import java.util.List;
 
 @RestController
-@RequestMapping(value = "/planetas")
+@RequestMapping(value = "/api")
 @Validated
 public class PlanetController {
 
@@ -41,29 +44,37 @@ public class PlanetController {
     // Busca planeta na base de dados pelo id e o número de filmes na Swapi
     @GetMapping(value = "/{id}")
     ResponseEntity<Planet> findPlanetById(@PathVariable String id) {
+        if (id.isBlank()) {
+            throw new InvalidIdException();
+        }
         return planetService.findPlanetById(id).map(record -> {
             PlanetSwapiPaginationResponse planetsApi = clientFeignSwapi.getPlanetsByName(record.getName());
             planetsApi.getResults().stream().forEach(pa -> {
                 record.setFilmsAppearances(pa.getFilms().size() + record.getFilmsAppearances());
             });
             return ResponseEntity.ok().body(record);
-        }).orElse(ResponseEntity.notFound().build());
+        }).orElseThrow(() -> new PlanetNotFoundException(id));
     }
+
 
     // Busca planeta por nome na base dados o número de filmes na Swapi
     @GetMapping(value = "/find/{name}")
     public List<Planet> findPlanetByName(@PathVariable String name, HttpServletResponse response) {
+        if (name.isBlank()) {
+            throw new InvalidNameException();
+        }
         List<Planet> planetList = planetService.findByNameIgnoreCase(name);
         planetList.stream().forEach(p -> {
             PlanetSwapiPaginationResponse planetsApi = clientFeignSwapi.getPlanetsByName(p.getName());
             planetsApi.getResults().stream().forEach(pa -> {
                 p.setFilmsAppearances(pa.getFilms().size() + p.getFilmsAppearances());
             });
+
         });
         if (!planetList.isEmpty()) {
             response.setStatus(HttpServletResponse.SC_OK);
         } else {
-            response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+            throw new NotFoundNameException(name);
         }
         return planetList;
     }
@@ -75,14 +86,14 @@ public class PlanetController {
         response.setStatus(HttpServletResponse.SC_CREATED);
     }
 
-    // Deleta um planeta da base de dados
+    // Deleta um planeta da base de dados pelo id
     @DeleteMapping(value = "/{id}")
     public void deletePlanet(@PathVariable String id, HttpServletResponse response) {
-       // if (id.equals(planetService.findPlanetById(id))) {
-             planetService.deletePlanet(id);
+        if (!id.isBlank()) {
+            planetService.deletePlanet(id);
             response.setStatus((HttpServletResponse.SC_OK));
-       // } else {
-       //     response.setStatus((HttpServletResponse.SC_NOT_FOUND));
-      //  }
+        } else {
+            throw new InvalidIdException();
+        }
     }
 }
